api/requirements.txt
reportlab==4.0.4
pandas==1.5.3
openpyxl==3.1.2
PyYAML==6.0.1